requirements.txt
alembic==1.7.7
auth0-python==4.7.0
databases==0.6.1
fastapi==0.109.0
fastapi_utils==0.2.1
loguru==0.6.0
pre-commit==3.6.0
pydantic[email]==1.10.2
pyjwt==2.8.0
pytest==7.1.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.31.0
sqlalchemy==1.4.41
sqlalchemy_utils==0.38.3
uvicorn==0.18.3
